# static/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API to interact with [Wakapi](https://wakapi.dev)\n\n## Authentication\nSet header `Authorization` to your API Key encoded as Base64 and prefixed with `Basic`\n**Example:** `Basic ODY2NDhkNzQtMTljNS00NTJiLWJhMDEtZmIzZWM3MGQ0YzJmCg==`",
        "title": "Wakapi API",
        "contact": {
            "name": "Ferdinand Mütsch",
            "url": "https://github.com/muety",
            "email": "ferdinand@muetsch.io"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://github.com/muety/wakapi/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/api/compat/shields/v1/{user}/{interval}/{filter}": {
            "get": {
                "description": "Retrieve total time for a given entity (e.g. a project) within a given range (e.g. one week) in a format compatible with [Shields.io](https://shields.io/endpoint). Requires public data access to be allowed.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "badges"
                ],
                "summary": "Get badge data",
                "operationId": "get-badge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to fetch data for",
                        "name": "user",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "today",
                            "yesterday",
                            "week",
                            "month",
                            "year",
                            "7_days",
                            "last_7_days",
                            "30_days",
                            "last_30_days",
                            "12_months",
                            "last_12_months",
                            "any"
                        ],
                        "type": "string",
                        "description": "Interval to aggregate data for",
                        "name": "interval",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter to apply (e.g. 'project:wakapi' or 'language:Go')",
                        "name": "filter",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.BadgeData"
                        }
                    }
                }
            }
        },
...
        "/api/compat/wakatime/v1/users/{user}/all_time_since_today": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mimics https://wakatime.com/developers#all_time_since_today",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wakatime"
                ],
                "summary": "Retrieve summary for all time",
                "operationId": "get-all-time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to fetch data for (or 'current')",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.AllTimeViewModel"
                        }
                    }
                }
            }
        },
...
        "/api/compat/wakatime/v1/users/{user}/projects": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mimics https://wakatime.com/developers#projects",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wakatime"
                ],
                "summary": "Retrieve and fitler the user's projects",
                "operationId": "get-wakatime-projects",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to fetch data for (or 'current')",
                        "name": "user",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Query to filter projects by",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ProjectsViewModel"
                        }
                    }
                }
            }
        },
...
        "/api/summary": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "summary"
                ],
                "summary": "Retrieve a summary",
                "operationId": "get-summary",
                "parameters": [
                    {
                        "enum": [
                            "today",
                            "yesterday",
                            "week",
                            "month",
                            "year",
                            "7_days",
                            "last_7_days",
                            "30_days",
                            "last_30_days",
                            "12_months",
                            "last_12_months",
                            "any"
                        ],
                        "type": "string",
                        "description": "Interval identifier",
                        "name": "interval",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (e.g. '2021-02-07')",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (e.g. '2021-02-08')",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Whether to recompute the summary from raw heartbeat or use cache",
                        "name": "recompute",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Summary"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Summary": {
            "type": "object",
            "properties": {
                "editors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "from": {
                    "type": "string",
                    "format": "date",
                    "example": "2006-01-02 15:04:05.000"
                },
                "labels": {
                    "description": "labels are not persisted, but calculated at runtime, i.e. when summary is retrieved",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "machines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "operating_systems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "to": {
                    "type": "string",
                    "format": "date",
                    "example": "2006-01-02 15:04:05.000"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.SummaryItem": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "v1.AllTimeData": {
            "type": "object",
            "properties": {
                "is_up_to_date": {
                    "description": "true if the stats are up to date; when false, a 202 response code is returned and stats will be refreshed soon\u003e",
                    "type": "boolean"
                },
                "range": {
                    "$ref": "#/definitions/v1.AllTimeRange"
                },
                "text": {
                    "description": "total time logged since account created as human readable string\u003e",
                    "type": "string"
                },
                "total_seconds": {
                    "description": "total number of seconds logged since account created",
                    "type": "number"
                }
            }
        },
        "v1.AllTimeRange": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "v1.AllTimeViewModel": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.AllTimeData"
                }
            }
        },
        "v1.BadgeData": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "schemaVersion": {
                    "type": "integer"
                }
            }
        },
        "v1.Project": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "repository": {
                    "type": "string"
                }
            }
        },
        "v1.ProjectsViewModel": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.Project"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}